Time: 118 ms (5.25%), Space: 89.4 MB (44.2%) - LeetHub

+++ 2358-number-of-ways-to-split-array/2358-number-of-ways-to-split-array.cpp
@@ -4,7 +4,7 @@ public:
 
       int ans = 0;
 
-      long int s = 0, e = 0;
+      long long int s = 0, e = 0;
 
       int n = nums.size()  ;
